Adds initial font and background color styling
This commit adds background color and font styling to the divs. The
styling is mostly what the page will end up with, although there will
be some further refinements and sub-divisions of the divs. I mostly
wanted to help differentiate the divs from each other while working.

This also adds font-family to the body style.