Changes organizaton of CSS file
This commit reorganizes the CSS file to have a separate section for
each of the 6 main div's in the HTML file. Each section in the CSS
file is noted with a comment. The intent is to have all styling for
each of the 6 div's in its own section. This commit also breaks up a
couple of shared styles (i.e., the same style applied to multiple
classes), which results in identical styles being applied to separate
classes in separate sections of the CSS. This is necessary in order
to organize the CSS file into separate sections for each of the main
div's.