Limits content width to 1000px
Commit to better align with the template image from The Odin Project
(https://cdn.statically.io/gh/TheOdinProject/curriculum
        /81a5d553f4073e593d23a6ab00d50eef8620796d/foundations
        /html_css/project/imgs/01.png)

The template image shows the actual content of the various sections
not going past about 1000px in width, having consistent left and
right margins between sections, and having the background colors
extend to the edges of the page. This could be accomplished with
sufficently large padding, but that did not seem like a good option.

Instead, this commit adds a common "sectionWrapper" classed div
inside each of the section div's. The individual section divs will
now be used only for setting background color. The sectionWrapper div
is set in the Common Styles section of CSS to be centered (margin:
auto) with a max-width of 1000px. Additionally, the display: flex is
moved into the Common Styles for sectionWrapper, as this class will
hold all of the HTML elements in each section. Section-level flex
properties will be set using the child or descendent combinators with
section div class as parent and sectionWrapper as child.