Enables top/bottom padding styles in main sections
To keep consistent side margins, padding was defined in CSS in
'body > div', which targets the main div of each of the main sections
of the page. However, prior to this commit, the margins were set with
'padding: 0px 24px;' in the body > div style. This style definition
has higer priority than a class style definition, meaning it cannot
be overwritten in a class' style. This is good for the side margins,
as I want to keep those consistent throughout the page so want to
prevent overwriting those, but I would like to be able to adjust the
top and bottom padding for each individual class. This commit
addresses this by replacing the above padding style with individual
'padding-left' and 'padding-right' styles in body > div.