Eliminates right margin on header links
Prior to this commit, li elements in the header had 10px margins on
all sides. This caused them not to align flush to the right margin of
the 1000px wide content section in the page. This commit changes the
margin on these li elements to be margin-left, so they are spaced
from each other but are aligned along the right margin of the
conent.