Fixes centering on Info section header text
Prior to this commit, CSS relied on column Flexbox and align-items
center to center the header text in the Info section. However, when
the window was shrunk horizontally to the point that the text needed
two lines, the centering was lost and text became left justified.
Added 'text-align: center' to the div style to address this issue.